Add isEmpty and containsSubscription(forKey:) to KeyedSubscriptionStore (#20)
- Add `isEmpty` to `SingleSubscriptionStore` and `KeyedSubscriptionStore`
- Add `containsSubscription(forKey:)` to `KeyedSubscriptionStore`

Be aware that these methods are thread-safe to use _(in that they won't cause a crash if called from multiple threads at the same time)_ but, by the time the results have been returned, the subscriptions inside of the store may have changed and the returned value won't be valid anymore. Therefore, use them with caution.